fix(ai): route custom Ark providers through endpoint-aware parsing (fixes #18540)
Per #18608 review feedback:

- Sink SUPPORTED_ENDPOINT_LIST into @shared/utils/api (format.ts/index.ts)
- Update hasApiVersion to recognize endpoint suffixes (B2)
- Reuse routeToEndpoint in listModels.ts to fix 404 on full endpoint URLs (B7-main)
- Fix renderer preview duplication by splitting baseURL/endpoint first (B7-renderer)
- provider.ts now imports canonical list instead of local copy (A1)
