+++ src/main/ai/provider/listModels.ts
@@ -35,7 +35,7 @@ import {
 import { SystemProviderIds } from '@shared/utils/systemProviderId'
 import * as z from 'zod'
 
-import { defaultHeaders, getBaseUrl } from '../utils/provider'
+import { defaultHeaders, getBaseUrl, routeToEndpoint } from '../utils/provider'
 import { COPILOT_DEFAULT_HEADERS } from './constants'
 import {
   createVertexModelListRequest,
@@ -720,7 +720,10 @@ const openAIFetcher: ModelFetcher = {
 const openAICompatibleFetcher: ModelFetcher = {
   match: () => true,
   fetch: async (provider, signal) => {
-    const baseUrl = formatApiHost(getBaseUrl(provider))
+    const rawBaseUrl = getBaseUrl(provider)
+    const { baseURL: host } = routeToEndpoint(rawBaseUrl)
+    const baseUrl = formatApiHost(host)
+
     const response = await getFromApi({
       url: `${baseUrl}/models`,
       headers: defaultHeaders(provider),
