+++ src/shared/utils/api/index.ts
@@ -9,6 +9,7 @@ export {
   joinApiKeyString,
   maskApiKey,
   splitApiKeyString,
+  SUPPORTED_ENDPOINT_LIST,
   withoutTrailingApiVersion,
   withoutTrailingSharp,
   withoutTrailingSlash
